Add new border option 'style'
This is a first try to allow styling your containers with CSS-like
properties.  Reading style sheets from file is not implemented yet.
Therefore to test this you will need to set some styles in the code.

The only properties that are currently supported are:
- border_width
- border_color
- border_radius
- background_color
- padding

Margin conflicts somehow with the 'inner gaps' option from sway.

But I realized that the wlr_renderer's capabilities aren't sufficient,
mainly because there is no way to render to a texture, meaning I would
have to render all decorations without hardware acceleration.  This
would be acceptable, if I wasn't planning for animating stuff.