Reimplement rendering of padding and borders
The new implementation draws borders and paddings in a single draw call
by rendering a single mesh, just like the rendering of shadows.  The
ability to draw rounded corners (border-radius) was dropped because it
simplifies things a lot and the previous implementation was buggy
anyway.  But I'm still planning to add this feature again.

This patch introduces a lot of duplicated code for shadow and border
rendering, which I will remove in a later commit when the final design
becomes clearer.