Implement rounded borders
To draw rounded borders the rendering is now done in two passes.  First
the border box is rendered with the border color and second the
background is rendered on top of that with the size of the padding box.
This removes a lot of edge cases.

Blurring edges with a guassian lookup table was also removed.  Instead
smoothstep() is used in the shader for blurring.  This makes it possible
to have blurring independently from the border radius.

This patch also removes a great amount of duplicated code by creating
style_render_decoration().